requirements.txt
crewai
langsmith
langchain
ollama
transformers
jinja2
python-dotenv
